+++ components/home/AboutUsSection.jsx
@@ -1,6 +1,6 @@
 import Image from 'next/image';
 import Button from '@/components/Button';
-import styles from '@/styles/home/Home.module.css';
+import styles from '@/styles/home/HomePage.module.css';
 
 const teams = [
 	{
